Can delete own player characters from a game now
List of PC will also show in the overview for a game now with the coupled player.
People can delete their character from a playthrough by going to the join url again and deleting their character.
There is now also styling coherent with the rest of the page on the create PC page.