added icon, title, reskin for login, and base skeleton for logs
The site now has an icon
The site now has a title
Reskinned login page to have icons for all inputs as well as a background image. Now also shows error messages on login and register when trying to do that.
Moved required images for the site to images/base/
Split main css into seperate css for login page for readability.